User Auth
API calls appropriately check user authorization.

+++ db/usersDelete.db.js
@@ -2,7 +2,7 @@ const mysql = require('mysql')
 const query = require('./query')
 
 const del = async (userId) => {
-  let sql = mysql.format('DELETE FROM users WHERE (id=?)', [userId])
+  let sql = mysql.format('DELETE FROM users WHERE (id=?) AND uid IS NULL', [userId])
   return await query(sql)
 }
